fix: receiving module REST API

+++ src/main/java/id/putra/wms/module/inbound/service/adapter/command/impl/PurchaseOrderCommandAdapterImpl.java
@@ -1,5 +1,7 @@
 package id.putra.wms.module.inbound.service.adapter.command.impl;
 
+import java.util.Objects;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -7,7 +9,6 @@ import id.putra.wms.module.inbound.dto.PurchaseOrderDto;
 import id.putra.wms.module.inbound.mapper.PurchaseOrderMapper;
 import id.putra.wms.module.inbound.model.repository.PurchaseOrderRepository;
 import id.putra.wms.module.inbound.service.adapter.command.PurchaseOrderCommandAdapter;
-import id.putra.wms.shared.enums.OrderStatus;
 import id.putra.wms.shared.enums.ResponseEnum;
 import id.putra.wms.shared.exceptions.ModuleException;
 import lombok.RequiredArgsConstructor;
@@ -23,17 +24,20 @@ public class PurchaseOrderCommandAdapterImpl implements PurchaseOrderCommandAdap
     @Override
     public PurchaseOrderDto add(PurchaseOrderDto dto) {
         var entity = purchaseOrderMapper.toEntity(dto);
-        return purchaseOrderMapper.toDto(purchaseOrderRepository.save(java.util.Objects.requireNonNull(entity)));
+        return purchaseOrderMapper.toDto(purchaseOrderRepository.save(Objects.requireNonNull(entity)));
     }
 
     @Override
     public PurchaseOrderDto update(PurchaseOrderDto dto) {
         var entity = purchaseOrderMapper.toEntity(dto);
-        return purchaseOrderMapper.toDto(purchaseOrderRepository.save(java.util.Objects.requireNonNull(entity)));
+        return purchaseOrderMapper.toDto(purchaseOrderRepository.save(Objects.requireNonNull(entity)));
     }
 
     @Override
     public Boolean delete(Long id) {
+        if (id == null) {
+            throw new ModuleException(ResponseEnum.INVALID_PARAM);
+        }
         if (purchaseOrderRepository.existsById(id)) {
             purchaseOrderRepository.deleteById(id);
             return true;
@@ -42,10 +46,10 @@ public class PurchaseOrderCommandAdapterImpl implements PurchaseOrderCommandAdap
     }
 
     @Override
-    public PurchaseOrderDto updateStatus(Long id, OrderStatus status) {
-        var entity = purchaseOrderRepository.findById(id)
+    public PurchaseOrderDto patch(PurchaseOrderDto dto) {
+        var entity = purchaseOrderRepository.findById(Objects.requireNonNull(dto.getId()))
                 .orElseThrow(() -> new ModuleException(ResponseEnum.DATA_NOT_FOUND));
-        entity.setStatus(status);
-        return purchaseOrderMapper.toDto(purchaseOrderRepository.save(entity));
+        purchaseOrderMapper.patchDtoToEntity(dto, entity);
+        return purchaseOrderMapper.toDto(purchaseOrderRepository.save(Objects.requireNonNull(entity)));
     }
 }
